+++ backend/swagger/docs/docs.go
@@ -2085,7 +2085,7 @@ const docTemplate = `{
                     "type": "integer"
                 },
                 "numImageProcessors": {
-                    "description": "number of concurrent image processing jobs",
+                    "description": "number of concurrent image processing jobs used to create previews, default is 4,",
                     "type": "integer"
                 },
                 "port": {